six_hats/requirements.txt
asgiref==3.2.3
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
cairocffi==1.0.2
CairoSVG==2.4.1
captcha==0.3
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cmake==3.13.3
constantly==15.1.0
cryptography==2.8
cssselect2==0.2.1
dataclasses==0.6
defusedxml==0.6.0
Django==2.2.2
django-allauth==0.39.1
django-common-helpers==0.9.2
django-cron==0.5.1
django-rest-auth==0.9.5
django-session-security==2.6.2
djangorestframework==3.10.2
docopt==0.6.2
docx==0.2.4
html5lib==1.0.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==4.8.1
incremental==17.5.0
lxml==4.3.2
Markdown==3.3.4
num2words==0.5.9
oauthlib==3.0.2
Pillow==6.1.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pycparser==2.19
PyHamcrest==1.9.0
PyPDF2==1.26.0
Pyphen==0.9.5
PyQRCode==1.2.1
python-dateutil==2.8.0
python-docx==0.8.10
python-dotenv==0.19.0
python3-openid==3.1.0
pytz==2019.1
reportlab==3.5.13
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.2.4
stripe==2.60.0
tinycss2==1.0.2
txaio==18.8.1
typing-extensions==3.10.0.2
urllib3==1.25.3
WeasyPrint==48
webencodings==0.5.1
xhtml2pdf==0.2.3
xlwt==1.3.0
zipp==3.6.0
zope.interface==4.7.1
